requirements.txt
fastapi~=0.103.2
pydantic~=2.4.2
uvicorn[standard]~=0.23.2
SQLAlchemy~=2.0.21
psycopg2-binary~=2.9.8
python-dotenv~=1.0.0
asyncpg~=0.28.0
redis~=5.0.1
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.6